pyproject.toml
[project]
name = "student-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.121.0",
    "joblib>=1.5.2",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.3",
    "pymongo>=4.15.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "uvicorn[standard]>=0.38.0",
]
